// src/components/confing/enums.ts
export enum templateHTML {
  template = '.goods__template',
  title = '.goods__title',
  img = '.goods__img',
  description = '.goods__description',
  price = '.goods__price',
  brand = '.goods__brand',
  year = '.goods__year',
  size = '.goods__size',
  shelf = '.goods__shelf',
  values = '.goods__value',
  cart = '.goods__cart',
  color = '.goods__color',
  link = '.goods__link',
  root = '.goods__list',
  buttons = '.goods__buttons',
  item = '.goods__item',
  notification = '.goods__notification',
  active = 'goods__item--active',
}

export enum Buttons {
  add = '.goods__button--add',
  remove = '.goods__button--remove',
  info = '.goods__button--info',
  storage = '.options__item--reset',
  filter = '.options__item--remove',
}

export enum Categories {
  root = 'categories__items',
  active = 'categories__item--active',
}

export enum Selection {
  root = '.sort__selection',
  option = 'option',
}

export enum SearchForm {
  root = '#search',
  notification = '.form__notification',
}

export enum CartElement {
  items = '.cart__items',
}

export enum Display {
  none = 'display--none',
}

export enum TargetFilterRange {
  price = 'price-range',
  year = 'year-range',
  value = 'value-range',
}

export enum TargetFilterType {
  brand = '.brand-filter__label',
  color = '.color-filter__label',
  checked = 'checked',
}